Cleanup with nvbench helpers
